requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
psycopg2-binary>=2.9
Flask-JWT-Extended>=4.0
python-dotenv>=1.0
Flask-Cors>=3.0
Werkzeug>=2.2
